python/requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.4.0
requests==2.32.5
urllib3==2.5.0
Werkzeug==3.0.3
PyJWT==2.8.0
bcrypt==4.1.2
